refactor move test modules
