+++ src/main/python/covata/delta/api/__init__.py
@@ -12,7 +12,7 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from interface import ApiClient
+from api_client import ApiClient
 from requests_client import RequestsApiClient
 
 __all__ = ['ApiClient', 'RequestsApiClient']
